persistence

+++ src/business/data/Position.java
@@ -1,30 +1,30 @@
 package business.data;
 
 public class Position {
-	private float x;
-	private float y;
+	private double x;
+	private double y;
 
-	public Position(float f, float g) {
+	public Position(double f, double g) {
 		this.x = f;
 		this.y = g;
 	}
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
-	public void Position(float x , float y) {
+	public void Position(double x , double y) {
 		this.x = x;
 		this.y = y;
 	}
